finish 2.9.7-fix 修复自动更新bug;升级所有金额字段为bigint;修复父科目展示问题

+++ eladmin-system/src/main/java/me/zhengjie/modules/keyuan/domain/SysProjectAccount.java
@@ -59,7 +59,7 @@ public class SysProjectAccount implements Serializable {
     @Column(name = "`initial_amount`", nullable = false)
     @NotNull
     @ApiModelProperty(value = "初始余额")
-    private Integer initialAmount;
+    private Long initialAmount;
 
     @Column(name = "`has_children`", nullable = false)
     @NotNull
